Fixed code with camera

+++ Logic/Entities/AbstractFactory.h
@@ -15,11 +15,11 @@ namespace logic {
     public:
         virtual ~AbstractFactory() = default;
 
-        virtual std::shared_ptr<EntityModel> createPacMan(double x, double y) = 0;
-        // virtual std::shared_ptr<EntityModel> createWall(double x, double y) = 0;
-        // virtual std::shared_ptr<EntityModel> createGhost(double x, double y) = 0;
-        // virtual std::shared_ptr<EntityModel> createCoin(double x, double y) = 0;
-        // virtual std::shared_ptr<EntityModel> createFruit(double x, double y) = 0;
+        virtual std::shared_ptr<EntityModel> createPacMan(float x, float y) = 0;
+        // virtual std::shared_ptr<EntityModel> createWall(float x, float y) = 0;
+        // virtual std::shared_ptr<EntityModel> createGhost(float x, float y) = 0;
+        // virtual std::shared_ptr<EntityModel> createCoin(float x, float y) = 0;
+        // virtual std::shared_ptr<EntityModel> createFruit(float x, float y) = 0;
     };
 }
